+++ backend/resolvers/index.ts
@@ -1,0 +1,6 @@
+//@ts-ignore
+import * as Query from "./query.ts";
+//@ts-ignore
+import * as Mutation from "./mutation.ts";
+
+export default { Query, Mutation };
