+++ backend/resolvers/types.ts
@@ -1,0 +1,12 @@
+export interface Success {
+  success: Boolean;
+}
+
+export type Transaction = {
+  amount: Number;
+  category: [String];
+  date: String;
+  merchant_name: String;
+  hidden: Boolean;
+  _id: String;
+};
