+++ tsconfig.json
@@ -13,5 +13,5 @@
     "isolatedModules": true,
     "lib": ["es5", "dom", "es2022"]
   },
-  "files": ["backend.ts"]
+  "files": ["backend/index.ts"]
 }
